* 去掉绝大多数的变异警告 * 格式化代码

+++ MPlayerEngine/Media.h
@@ -2,22 +2,22 @@
 
 #include "MPTime.h"
 
-class CMedia : public IMedia
-{
+
+class CMedia : public IMedia {
     OBJ_REFERENCE_DECL
 public:
     CMedia();
     virtual ~CMedia();
 
     // the ID in media library, 0 for not exist in media library
-    virtual long getID();
+    virtual int getID();
 
     virtual MLRESULT getSourceUrl(IString *strUrl);
     virtual MLRESULT getArtist(IString *strArtist);
     virtual MLRESULT getTitle(IString *strTitle);
     virtual MLRESULT getAlbum(IString *strAlbum);
 
-    virtual long getDuration();
+    virtual int getDuration();
 
     virtual MLRESULT setSourceUrl(cstr_t strUrl);
 
@@ -36,38 +36,38 @@ public:
 protected:
     friend class CMediaLibrary;
 
-    std::mutex                m_mutexDataAccess;
+    std::mutex                  m_mutexDataAccess;
 
-    long                m_nID;
+    int                         m_nID;
 
-    string                m_strUrl;
-    string                m_strArtist;
-    string                m_strAlbum;
-    string                m_strTitle;
-    int16_t                m_nTrackNumb;
-    int16_t                m_nYear;
-    string                m_strGenre;
-    string                m_strComment;
-    long                m_nLength;
-    long                m_nFileSize;
-    CMPTime                m_nTimeAdded;
-    CMPTime                m_nTimePlayed;
+    string                      m_strUrl;
+    string                      m_strArtist;
+    string                      m_strAlbum;
+    string                      m_strTitle;
+    int16_t                     m_nTrackNumb;
+    int16_t                     m_nYear;
+    string                      m_strGenre;
+    string                      m_strComment;
+    int                         m_nLength;
+    uint32_t                    m_nFileSize;
+    CMPTime                     m_nTimeAdded;
+    CMPTime                     m_nTimePlayed;
 
-    uint16_t                m_nRating;    // 0 ~ 500
-    bool                m_bIsUserRating;
-    bool                m_bIsFileDeleted;
-    long                m_nPlayed;
-    long                m_nPlaySkipped;
+    uint16_t                    m_nRating;          // 0 ~ 500
+    bool                        m_bIsUserRating;
+    bool                        m_bIsFileDeleted;
+    int                         m_nPlayed;
+    int                         m_nPlaySkipped;
 
-    string                m_strLyricsFile;
+    string                      m_strLyricsFile;
 
-    string                m_strFormat;
-    long                m_nBitRate;
-    bool                m_bIsVbr;
-    uint8_t                m_nChannels;
-    uint8_t                m_nBPS;
-    bool                m_bInfoUpdated;        // if InfoUpdated, it will be update to database when playing finished.
-    uint32_t                m_nSampleRate;
-    string                m_strExtraInfo;
+    string                      m_strFormat;
+    int                         m_nBitRate;
+    bool                        m_bIsVbr;
+    uint8_t                     m_nChannels;
+    uint8_t                     m_nBPS;
+    bool                        m_bInfoUpdated;     // if InfoUpdated, it will be update to database when playing finished.
+    uint32_t                    m_nSampleRate;
+    string                      m_strExtraInfo;
 
 };
